Create Transfer_2_SimpleData.C.
The original data format of SpiRITROOT include all the information in the track reconstruction. It is quite large and slow when you load these information. So here I just select some information that I think it is useful in the data analysis. Up to now, in the data analysis, almost each person will have his only similar function code, so merging this script to the SpiRITROOT will make it messy.